Change %s thingies once more
git-svn-id: https://ecesvn.uwaterloo.ca/courses/ece108/w18/students/mdcbonne/SatSolver@2404 bf7adb0c-0a71-4f8a-9a9b-06af373ef5e1

+++ src/main/main.c
@@ -98,10 +98,10 @@ char *classify() {
 			truePermutations++;
 		}
 	}
-	printf("\nNumber of permutations: %d True Permutations: %d\n", numPermutations, truePermutations);
+	printf("\nNumber of permutations: %u True Permutations: %f\n", numPermutations, truePermutations);
 	if (truePermutations == 0){
 		return "unsatisfiable";
-	} else if (truePermutations - 1 == numPermutations){
+	} else if ((double)truePermutations - 1 == (double)numPermutations){
 		return "tautology";
 	} else {
 		return "satisfiable";
